corrections after review: openPopup moved to utils, added missing placeholders
