+++ src/index.html
@@ -36,11 +36,11 @@
       <h2 class="popup__title">Редактировать профиль</h2>
       <form class="popup__form" name="profile-edit-form" id='profile-edit-form'>
         <label for="profile-name-input" class="popup__field">
-          <input type="text" class="popup__input-text popup__input-text_field_name" id="profile-name-input" name="profile-name-input" value="Жак-Ив Кусто" required minlength=2 maxlength=40>
+          <input type="text" class="popup__input-text popup__input-text_field_name" id="profile-name-input" name="profile-name-input" value="Жак-Ив Кусто" placeholder="Имя пользователя" required minlength=2 maxlength=40>
           <span class="popup__input-error profile-name-input-error">test</span>
         </label>
         <label for="profile-bio-input" class="popup__field">
-          <input type="text" class="popup__input-text popup__input-text_field_bio" id="profile-bio-input" name="profile-bio-input" value="Исследователь океана" required minlength=2 maxlength=200>
+          <input type="text" class="popup__input-text popup__input-text_field_bio" id="profile-bio-input" name="profile-bio-input" value="Исследователь океана" placeholder="О себе" required minlength=2 maxlength=200>
           <span class="popup__input-error profile-bio-input-error">test</span>
         </label>
         <button class="popup__save" type="submit">Сохранить</button>
